Extract HomePage and EnglishHomePage page components
