--- components/home/index.ts
@@ -1,5 +1,0 @@
-export { default as Hero } from "./hero";
-export { default as OurValues } from "./our-values";
-export { default as Numbers } from "./numbers";
-export { default as ImageGallery } from "./image-gallery/index";
-export { default as Partners } from "./partners";
