+++ pages/en.tsx
@@ -1,51 +1,17 @@
 import { NextPage, GetStaticProps } from "next";
 import { PortalPartner } from "lib/portal-types";
-import { Layout, Section, SectionContent } from "components/layout";
-import { ThemeContext } from "styled-components";
-import { useContext } from "react";
-import {
-  Hero,
-  OurValues,
-  Numbers,
-  ImageGallery,
-  Partners,
-} from "components/home";
 import { siteData } from "lib/site-data";
+import EnglishHomePage from "components/home/english";
 
 type PageProps = {
   partners: PortalPartner[];
 };
 
 const Page: NextPage<PageProps> = ({ partners }) => {
-  const theme = useContext(ThemeContext);
-  return (
-    <Layout lang="en">
-      <Section>
-        <Hero lang="en" />
-      </Section>
-
-      <Numbers lang="en" />
-
-      <Section backgroundColor={theme.colors.pebble}>
-        <SectionContent>
-          <OurValues lang="en" />
-        </SectionContent>
-      </Section>
-
-      <Section>
-        <ImageGallery />
-      </Section>
-
-      <Section>
-        <SectionContent>
-          <Partners partners={partners} lang="en" />
-        </SectionContent>
-      </Section>
-    </Layout>
-  );
+  return <EnglishHomePage partners={partners} />;
 };
 
-export const getStaticProps: GetStaticProps<PageProps> = async (context) => {
+export const getStaticProps: GetStaticProps<PageProps> = async () => {
   const homepagePartners = siteData.partners.filter((p) =>
     p.categories.some((c) => c === "homepage")
   );
