+++ pages/index.tsx
@@ -1,22 +1,9 @@
 import type { NextPage, GetStaticProps } from "next";
-import { useRouter } from "next/router";
-import { useContext } from "react";
-
 import { PortalPartner, PortalProject } from "lib/portal-types";
 import { siteData } from "lib/site-data";
-
-import { Layout, Section, SectionContent } from "components/layout";
-import { Projects, JoinUs } from "components/sections";
-import { ThemeContext } from "styled-components";
-
-import {
-  Hero,
-  OurValues,
-  Numbers,
-  ImageGallery,
-  Partners,
-} from "components/home";
 import { shuffleInPlace } from "lib/utils";
+import HomePage from "components/home";
+import { useRouter } from "next/router";
 
 type PageProps = {
   featuredProjects: readonly PortalProject[];
@@ -24,48 +11,14 @@ type PageProps = {
 };
 
 const Page: NextPage<PageProps> = ({ featuredProjects, partners }) => {
-  const theme = useContext(ThemeContext);
   const router = useRouter();
   const displayBanner = !!router.query.banner;
-
   return (
-    <Layout showBanner={displayBanner}>
-      <Section>
-        <Hero />
-      </Section>
-
-      <Numbers />
-
-      <Section>
-        <SectionContent>
-          <Projects projects={featuredProjects} />
-        </SectionContent>
-      </Section>
-
-      <Section>
-        <SectionContent>
-          <JoinUs />
-        </SectionContent>
-      </Section>
-
-      <Section backgroundColor={theme.colors.pebble}>
-        <SectionContent>
-          <OurValues />
-        </SectionContent>
-      </Section>
-
-      <Section>
-        <ImageGallery />
-      </Section>
-
-      {partners.length > 0 && (
-        <Section>
-          <SectionContent>
-            <Partners partners={partners} />
-          </SectionContent>
-        </Section>
-      )}
-    </Layout>
+    <HomePage
+      featuredProjects={featuredProjects}
+      partners={partners}
+      displayDonationBanner={displayBanner}
+    />
   );
 };
